Switch back to apex domain
Personal preference. Also, many of my blog inspirations who use their name as their personal domain are doing it. To name a few:

- https://github.com/tomcritchlow/tomcritchlow.github.io
- https://github.com/jingruzhang/jingruzhang.github.io
- https://github.com/swanson/swanson.github.com

For SEO, it doesn't matter which I choose, as long as I stick to it. No more switching.

Note: isaacs/github#1675 is still unresolved, so `https://www.joshjavier.com/` will not work, but it's a trade-off I'm willing to make.